move v1 c-api into a feature flag too
