+++ CONTRIBUTING.md
@@ -6,7 +6,7 @@ Thank you for your interest in contributing to duckdb-rs. This guide describes h
 
 This workspace has four crates:
 
-- `libduckdb-sys` provides native bindings for the [DuckDB C API](https://duckdb.org/docs/current/clients/c/api). With the `capi-v2` feature it also exposes bindings for the v2 C API (`duckdb_v2.h`) under `libduckdb_sys::v2`.
+- `libduckdb-sys` provides native bindings for the [DuckDB C API](https://duckdb.org/docs/current/clients/c/api) under `libduckdb_sys::v1` (default `capi-v1` feature, re-exported at the crate root). With the `capi-v2` feature it also exposes bindings for the v2 C API (`duckdb_v2.h`) under `libduckdb_sys::v2`.
 - `duckdb` provides an ergonomic wrapper around the v1 C API.
 - `duckdb-rs-neo` provides an ergonomic wrapper around the v2 C API. The v2 C API is still in rapid development, so this crate is experimental. Its `capi-v2-p2` feature targets API surface the bundled DuckDB does not provide yet and does not compile.
 - `duckdb-loadable-macros` provides procedural macros for loadable DuckDB extensions.
