+++ crates/duckdb-rs-neo/Cargo.toml
@@ -32,7 +32,10 @@ capi-v2-p2 = []
 
 [dependencies]
 libc = { workspace = true }
-libduckdb-sys = { workspace = true, features = ["capi-v2"] }
+# Not inherited from the workspace: cargo does not allow overriding
+# `default-features` on an inherited dependency. Only the v2 bindings are
+# needed; keep the library-discovery defaults. upgrade.sh bumps the version pin.
+libduckdb-sys = { version = "=1.20000.0", path = "../libduckdb-sys", default-features = false, features = ["capi-v2", "pkg-config", "vcpkg"] }
 paste = { workspace = true }
 r2d2 = { workspace = true, optional = true }
