+++ crates/libduckdb-sys/Cargo.toml
@@ -21,7 +21,7 @@ edition = { workspace = true }
 rust-version = { workspace = true }
 
 [features]
-default = ["vcpkg", "pkg-config"]
+default = ["capi-v1", "vcpkg", "pkg-config"]
 bundled = ["cc"]
 # Warning: experimental feature
 bundled-cmake = ["bundled", "dep:cmake", "dep:which", "parquet"]
@@ -35,7 +35,11 @@ tpch = ["bundled-cmake"]
 extensions-full = ["json", "parquet"]
 winduckdb = []
 # Warning: experimental feature
-loadable-extension = ["prettyplease", "quote", "syn"]
+loadable-extension = ["prettyplease", "quote", "syn", "capi-v1"]
+# Generate bindings for the v1 C API header (`duckdb.h`), exposed as `libduckdb_sys::v1`
+# and re-exported at the crate root. Disable with `default-features = false` when only
+# the v2 API is needed.
+capi-v1 = []
 # Also generate bindings for the v2 C API header (`duckdb_v2.h`), exposed as `libduckdb_sys::v2`.
 capi-v2 = []
